+++ lib/src/slice.dart
@@ -4,7 +4,7 @@ import 'sort.dart';
 
 /// A slice of data that indicates whether there's a next or previous slice available.
 /// Allows to obtain a [Pageable] to request a previous or next [Slice].
-abstract class Slice<T> extends Iterable<T> {
+abstract class Slice<T> {
   /// Returns the number of the current [Slice]. Is always non-negative.
   int get number;
 
@@ -54,9 +54,8 @@ abstract class Slice<T> extends Iterable<T> {
   /// first one.
   Pageable get previousOrFirstPageable => hasPrevious ? previousPageable : pageable;
 
-  /// Returns a new [Slice] with the content of the current one mapped by the given [toElement].
-  @override
-  Slice<U> map<U>(U Function(T e) toElement);
+  /// Returns a new [Slice] with the content of the current one mapped by the given [converter].
+  Slice<U> map<U>(U Function(T e) converter);
 }
 
 /// Abstract basis for [Slice] implementations.
@@ -100,11 +99,8 @@ abstract class Chunk<T> extends Slice<T> {
   @override
   Sort get sort => pageable.sort;
 
-  @override
-  Iterator<T> get iterator => content.iterator;
-
-  /// Applies the given [toElement] function to the content of this [Chunk].
-  List<U> getConvertedContent<U>(U Function(T e) toElement) => content.map(toElement).toList();
+  /// Applies the given [converter] function to the content of this [Chunk].
+  List<U> getConvertedContent<U>(U Function(T e) converter) => content.map(converter).toList();
 
   @override
   bool operator ==(Object other) {
@@ -135,8 +131,8 @@ class SliceImpl<T> extends Chunk<T> {
   SliceImpl({required super.content, super.pageable = Unpaged.unsorted, this.hasNext = false});
 
   @override
-  Slice<U> map<U>(U Function(T e) toElement) =>
-      SliceImpl<U>(content: getConvertedContent(toElement), pageable: pageable, hasNext: hasNext);
+  Slice<U> map<U>(U Function(T e) converter) =>
+      SliceImpl<U>(content: getConvertedContent(converter), pageable: pageable, hasNext: hasNext);
 
   @override
   bool operator ==(Object other) {